--- a/device/common/include/traccc/edm/device/sort_key.hpp
+++ b/device/common/include/traccc/edm/device/sort_key.hpp
@@ -9,6 +9,7 @@
 
 // Project include(s).
 #include "traccc/definitions/primitives.hpp"
+#include "traccc/edm/track_candidate.hpp"
 #include "traccc/edm/track_parameters.hpp"
 
 namespace traccc::device {
@@ -18,7 +19,7 @@ using sort_key = traccc::scalar;
 TRACCC_HOST_DEVICE
 inline sort_key get_sort_key(const bound_track_parameters& params) {
     // key = |theta - pi/2|
-    return math::abs(params.theta() - constant<traccc::scalar>::pi_2);
+    return math::fabs(params.theta() - constant<traccc::scalar>::pi_2);
 }
 
 }  // namespace traccc::device