--- a/device/sycl/src/fitting/fitting_algorithm.sycl
+++ b/device/sycl/src/fitting/fitting_algorithm.sycl
@@ -7,6 +7,7 @@
 
 // Project include(s).
 #include "../utils/get_queue.hpp"
+#include "traccc/fitting/device/fill_sort_keys.hpp"
 #include "traccc/fitting/device/fit.hpp"
 #include "traccc/fitting/kalman_filter/kalman_fitter.hpp"
 #include "traccc/sycl/fitting/fitting_algorithm.hpp"
@@ -18,6 +19,10 @@
 #include "detray/navigation/navigator.hpp"
 #include "detray/propagator/rk_stepper.hpp"
 
+// DPL include(s).
+#include <oneapi/dpl/algorithm>
+#include <oneapi/dpl/execution>
+
 // System include(s).
 #include <vector>
 
@@ -27,6 +32,9 @@ namespace kernels {
 /// Class identifying the kernel running @c
 /// traccc::device::fit
 class fit;
+/// Class identifying the kernel running @c
+/// traccc::device::fill_sort_keys
+class fill_sort_keys;
 }  // namespace kernels
 
 template <typename fitter_t>
@@ -74,19 +82,47 @@ track_state_container_types::buffer fitting_algorithm<fitter_t>::operator()(
     // (=32)
     unsigned int localSize = 64;
 
+    vecmem::data::vector_buffer<device::sort_key> keys_buffer(n_tracks,
+                                                              m_mr.main);
+    vecmem::data::vector_buffer<unsigned int> param_ids_buffer(n_tracks,
+                                                               m_mr.main);
+    vecmem::data::vector_view<device::sort_key> keys_view(keys_buffer);
+    vecmem::data::vector_view<unsigned int> param_ids_view(param_ids_buffer);
+
+    // Sort the key to get the sorted parameter ids
+    vecmem::device_vector<device::sort_key> keys_device(keys_buffer);
+    vecmem::device_vector<unsigned int> param_ids_device(param_ids_buffer);
+
     // 1 dim ND Range for the kernel
     auto trackParamsNdRange =
         traccc::sycl::calculate1DimNdRange(n_tracks, localSize);
 
     details::get_queue(m_queue)
         .submit([&](::sycl::handler& h) {
+            h.parallel_for<kernels::fill_sort_keys>(
+                trackParamsNdRange, [track_candidates_view, keys_view,
+                                     param_ids_view](::sycl::nd_item<1> item) {
+                    device::fill_sort_keys(item.get_global_linear_id(),
+                                           track_candidates_view, keys_view,
+                                           param_ids_view);
+                });
+        })
+        .wait_and_throw();
+
+    oneapi::dpl::sort_by_key(oneapi::dpl::execution::dpcpp_default,
+                             keys_device.begin(), keys_device.end(),
+                             param_ids_device.begin());
+
+    details::get_queue(m_queue)
+        .submit([&](::sycl::handler& h) {
             h.parallel_for<kernels::fit>(
                 trackParamsNdRange,
                 [det_view, field_view, config = m_cfg, track_candidates_view,
-                 track_states_view](::sycl::nd_item<1> item) {
-                    device::fit<fitter_t>(
-                        item.get_global_linear_id(), det_view, field_view,
-                        config, track_candidates_view, track_states_view);
+                 param_ids_view, track_states_view](::sycl::nd_item<1> item) {
+                    device::fit<fitter_t>(item.get_global_linear_id(), det_view,
+                                          field_view, config,
+                                          track_candidates_view, param_ids_view,
+                                          track_states_view);
                 });
         })
         .wait_and_throw();