--- a/extern/dpl/README.txt
+++ b/extern/dpl/README.txt
@@ -1,0 +1,4 @@
+# Build Recipe for oneDPL
+
+This directory holds a build recipe for building
+[DPL](https://github.com/oneapi-src/oneDPL) for this project.