Based on #706 

### UPDATE

Now the KF also uses the sorting method, but with a different standard: the number of measurements ($$n$$) of the track. 

The problem of using $$\theta$$ method is that the branching still can be an issue if the some of tracks in the warp end too early unexpectedly (or continue with too many measurements). We can prevent them by sorting the tracks w.r.t to the number of measurements.

One can still argue that the different step size between the planes can be problematic, and I think that is a valid point :thinking: 
It is not easy to know which one between $$\theta$$ or $$n$$ is better until one check the results. At least the ODD data says that sorting with $$n$$ is beneficial.

<google-sheets-html-origin><!--td {border: 1px solid #cccccc;}br {mso-data-placement:same-cell;}-->
  | Sort CKF and KF (PR) | Sort CKF | No sort (main)
-- | -- | -- | --
mt cuda [Hz] | 19.01 | 17.85 | 15.0808

Of course we can try to improve this further by experimenting different values or combining them

-------------------------------------------------------------------------


The main purpose of the PR is sorting the track parameters with based on $$\theta$$, to reduce the branching divergence. (Similar to the binning method that Markus mentioned in the parallelization meeting) 

As it turned out that sorting the track parameters themselves is quite expensive, I decided to make a small vector holding $$\theta$$ and vector indices, which is sorted instead. In the kernel, the track parameters are accessed by the indices of this small vector.

Following is the benchmark result with `traccc_throughput_mt_cuda` ODD data and `traccc_benchmark_cuda` with toy geometry

Command for **mt_cuda**: 
`./bin/traccc_throughput_mt_cuda --detector-file=geometries/odd/odd-detray_geometry_detray.json --digitization-file=geometries/odd/odd-digi-geometric-config.json --grid-file=geometries/odd/odd-detray_surface_grids_detray.json --use-detray-detector --input-directory=../../../../../../../bld6/data/traccc/geant4_ttbar_mu100/ --cpu-threads 2`

Command for **cuda toy benchmark**:
`./bin/traccc_benchmark_cuda`

Device: NVIDIA RTX A6000

![image](https://github.com/user-attachments/assets/e878344e-c2a2-4c49-ada9-04658a58c7e8)