@beomki-yeo 2024-09-25
The KF sort is done. As I need to use the sort function in SYCL KF, I included oneDPL in this PR. I guess this is OK as oneapi-src/oneDPL#1060 is resolved (@krasznaa). But I actually did not run the SYCL KF because the example executables are not being compiled at the moment (#655)

Following is the performance of this PR

This PR
```
Reconstructed track parameters: 3299852
Time totals:
                  File reading  635 ms
            Warm-up processing  578 ms
              Event processing  4864 ms
Throughput:
            Warm-up processing  57.807 ms/event, 17.2989 events/s
              Event processing  48.641 ms/event, 20.5588 events/s
```

Main
```
Reconstructed track parameters: 3299932
Time totals:
                  File reading  622 ms
            Warm-up processing  606 ms
              Event processing  5285 ms
Throughput:
            Warm-up processing  60.6916 ms/event, 16.4767 events/s
              Event processing  52.8595 ms/event, 18.9181 events/s
```