@beomki-yeo 2024-10-01
I checked the performance again and I found that this PR makes the `mt_cuda` slower.. No idea why I got the opposite result in the previous attempt.
 let me do the fix