@krasznaa 2024-10-01
Did you check the MT performance earlier? Because I could imagine situations where the ST performance would improve, but the MT performance would decrease.

In any case, as long as you have an idea what to do about it, I'll just let you propose an update. :wink: