@beomki-yeo 2024-10-01
I used the MT performance for the previous attempt as well. Well let's see..