@beomki-yeo 2024-10-01
I think I was sane. For the previous attempt, I configured cmake with `-DCMAKE_CUDA_ARCHITECTURES=86` for A6000 (it seems the correct number according to [here](https://arnon.dk/matching-sm-architectures-arch-and-gencode-for-various-nvidia-cards/)) and this literally makes an opposite outcome compared to the configuration without it. With the architecture number, the performance increases for $$n$$ sorting but drops for $$\theta$$. The behaviors are opposite with the other configuration.. 

<google-sheets-html-origin><!--td {border: 1px solid #cccccc;}br {mso-data-placement:same-cell;}-->
A6000, Release build |   |   |  
-- | -- | -- | --
Sort key | \|θ- π/2\| | Number of measurements ($$n$$) | No sort (Commit: 9b4c18db)
mt_cuda (w/ cmake arch num:86) [Hz] | 16.69 | 20.87 | 19.47
mt_cuda [Hz] | 21.3 | 20.19 | 20.787

@krasznaa Do you know what's going on here? Should I not use the CUDA architecture number for cmake configuration? The automatically configured architecture number is 52