@krasznaa 2024-10-01
I might be misunderstanding you, but what you found is that Thrust's sorting speed / efficiency is very optimization dependent, no? :confused:

When we sort the tracks in an extra step, we need the total runtime of the sorting step to be shorter than the amount by which the other steps speed up because of the sorting. I don't think that our own code would change its speed all too much because of which exact architecture we build for. So it must be the sorting that is way faster with SM 8.6 compared to SM 5.2.

The SM 5.2 value does not have any big significance. That's the minimum value that modern CUDA versions support without a warning. Normally I'd say that we should build the code for a couple of different architectures in parallel. The reason I didn't pursue this yet is because of our large memory usage as-is. :thinking: But yeah, as long as build system issues are sorted out, we could eventually go for `CMAKE_CUDA_ARCHITECTURES=all-major` in our build. See:

https://cmake.org/cmake/help/latest/prop_tgt/CUDA_ARCHITECTURES.html

What we could do for now: If we take the hit of raising our minimum CMake version to 3.24, we could set `CMAKE_CUDA_ARCHITECTURES=native` in our build. :thinking: That will make the code pretty unportable, but at this stage of development it may still be the best setup to have.