@beomki-yeo 2024-10-01
To be clear, the numbers in the table are the event throughput of `mt_cuda`, not the sorting speed.

> I might be misunderstanding you, but what you found is that Thrust's sorting speed / efficiency is very optimization dependent, no? 😕

Probably right but I am not sure if the odd behavior of table numbeers is fully dependent on thrust sort speed. As you can see the behavior of the first and second column is totally opposite. If it is really on the thrust sort speed, they would behave consistently. Most of all, the sort function is called only once in the KF (unlike CKF where it is called more than 10 times)

> So it must be the sorting that is way faster with SM 8.6 compared to SM 5.2.

Probably not. The first column way slower with SM 8.6. (But As I mentioned, I doubt that the thrust sort is the dominating factor.. considering the huge gap)

Using `Native` sounds good. Though I am keen to use the number which shows the peak performance for the specific device (yeah it is 5.2 for A6000 :thinking:...)