@krasznaa 2024-10-01
Ahh... so the throughput tests are slower with SM 8.6? :confused: That I didn't expect indeed. (And hence didn't even check the numbers, as this would seem like a very unlikely outcome.)

This would be worth investigating in more detail, with some profiling.

But then, do I understand correctly that for now we do nothing? Since the current setup, with SM 5.2, gives optimal performance in your tests apparently.  :thinking: