@beomki-yeo 2024-10-01
Yup, doing nothing at the moment is my suggestion. (and I will probably make a PR sorting with $$\theta - \pi/2$$.) More profiling will be needed definitely